+++ piGantry/piSerial/__init__.py
@@ -2,3 +2,4 @@
 from .serialInter import *
 from .TkInterUI import *
 from .motorFunc import *
+from .mathFunc import *