+++ CHANGELOG.md
@@ -7,7 +7,9 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 - Reworked piSerial serialInter.py
 - piSerial now functional module
 ### Removed
-
+- opencv2test.py
+- imageDiff.py
+- imageProcess redundant files 
 ### Added
 - cameraFunc.py for handling image processing functions
 - Added average.py for pixelwise alternative to opencv contouring